requirements.txt
fastapi==0.124.4
httpx==0.28.1
lxml==6.0.2
openpyxl==3.1.5
pandas==2.3.3
Pillow==12.0.0
pydantic_settings==2.12.0
Requests==2.32.5
uvicorn==0.38.0
xlrd==2.0.1
